+++ .idea/gradle.xml
@@ -9,8 +9,8 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/sentry-android" />
+            <option value="$PROJECT_DIR$/sentry-app" />
           </set>
         </option>
       </GradleProjectSettings>
