+++ .idea/libraries/appcompat_v7_23_1_1.xml
@@ -1,11 +1,11 @@
 <component name="libraryTable">
   <library name="appcompat-v7-23.1.1">
     <ANNOTATIONS>
-      <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.android.support/appcompat-v7/23.1.1/annotations.zip!/" />
+      <root url="jar://$PROJECT_DIR$/sentry-android/build/intermediates/exploded-aar/com.android.support/appcompat-v7/23.1.1/annotations.zip!/" />
     </ANNOTATIONS>
     <CLASSES>
-      <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.android.support/appcompat-v7/23.1.1/jars/classes.jar!/" />
-      <root url="file://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.android.support/appcompat-v7/23.1.1/res" />
+      <root url="jar://$PROJECT_DIR$/sentry-android/build/intermediates/exploded-aar/com.android.support/appcompat-v7/23.1.1/jars/classes.jar!/" />
+      <root url="file://$PROJECT_DIR$/sentry-android/build/intermediates/exploded-aar/com.android.support/appcompat-v7/23.1.1/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
