+++ .idea/modules.xml
@@ -3,8 +3,8 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/Sentry-Android.iml" filepath="$PROJECT_DIR$/Sentry-Android.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/sentry-android/sentry-android.iml" filepath="$PROJECT_DIR$/sentry-android/sentry-android.iml" />
+      <module fileurl="file://$PROJECT_DIR$/sentry-app/sentry-app.iml" filepath="$PROJECT_DIR$/sentry-app/sentry-app.iml" />
     </modules>
   </component>
 </project>
