+++ app/build.gradle
@@ -3,11 +3,9 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 23
     buildToolsVersion "23.0.2"
-
     useLibrary 'org.apache.http.legacy'
-
     defaultConfig {
-        applicationId "com.joshdholtz.sentry"
+        applicationId 'com.joshdholtz.sentrytesting'
         minSdkVersion 15
         targetSdkVersion 22
         versionCode 1
@@ -19,9 +17,12 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    productFlavors {
+    }
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile fileTree(include: ['*.jar'], dir: 'libs')
     compile 'com.android.support:appcompat-v7:23.1.1'
+    compile project(':sentry-android')
 }
